+++ .github/CODEOWNERS
@@ -1,4 +1,2 @@
-# Add a team or username to this file
-# Example:
-# *   @ministryofjustice/operations-engineering
-@PhilDigitalJustice @venkat980 @karlbaker02
+ *   @ministryofjustice/laa-data-stewardship-payments-team
+*    @ministryofjustice/payforlegalaid
