+++ .github/dependabot.yml
@@ -25,10 +25,10 @@ updates:
     groups:
       gradle-updates:
         exclude-patterns:
-          - "uk.gov.laa.springboot.microservice.*"
+          - "uk.gov.laa.dstew.claimsreports.*"
       internal-packages:
         patterns:
-          - "uk.gov.laa.springboot.microservice*"
+          - "uk.gov.laa.dstew.claimsreports*"
   - package-ecosystem: "github-actions"
     directory: "/"
     schedule:
