+++ .github/workflows/build-test-pr.yml
@@ -44,7 +44,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: checkstyle-report
-          path: spring-boot-microservice-service/build/reports/checkstyle
+          path: laa-data-claims-reporting-service/build/reports/checkstyle
           retention-days: 14
 
       - name: Upload test report
@@ -52,7 +52,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-report
-          path: spring-boot-microservice-service/build/reports/tests
+          path: laa-data-claims-reporting-service/build/reports/tests
           retention-days: 14
 
       - name: Upload jacoco coverage report
@@ -60,7 +60,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: jacoco-coverage-report
-          path: spring-boot-microservice-service/build/reports/jacoco
+          path: laa-data-claims-reporting-service/build/reports/jacoco
           retention-days: 14
 
   vulnerability-scan-app:
