+++ Dockerfile
@@ -6,7 +6,7 @@ RUN mkdir -p /opt/laa-spring-boot-microservice/
 WORKDIR /opt/laa-spring-boot-microservice/
 
 # Copy the JAR file into the container
-COPY spring-boot-microservice-service/build/libs/spring-boot-microservice-service-1.0.0.jar app.jar
+COPY laa-data-claims-reporting-service/build/libs/laa-data-claims-reporting-service-1.0.0.jar app.jar
 
 # Create a group and non-root user
 RUN addgroup -S appgroup && adduser -u 1001 -S appuser -G appgroup
