+++ build.gradle
@@ -1,3 +1,3 @@
 subprojects {
-    group = 'uk.gov.justice.laa.springboot.microservice'
+    group = 'uk.gov.justice.laa.dstew.claimsreports'
 }
