+++ laa-data-claims-reporting-service/lombok.config
