+++ laa-data-claims-reporting-service/src/main/java/uk/gov/justice/laa/dstew/claimsreports/SpringBootMicroserviceApplication.java
@@ -1,0 +1,20 @@
+package uk.gov.justice.laa.dstew.claimsreports;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Entry point for the Spring Boot microservice application.
+ */
+@SpringBootApplication
+public class SpringBootMicroserviceApplication {
+
+  /**
+   * The application main method.
+   *
+   * @param args the application arguments.
+   */
+  public static void main(String[] args) {
+    SpringApplication.run(SpringBootMicroserviceApplication.class, args);
+  }
+}
