+++ settings.gradle
@@ -14,13 +14,9 @@ pluginManagement {
 
     plugins {
         id 'uk.gov.laa.springboot.laa-spring-boot-gradle-plugin' version '1.0.0'
-        id 'org.openapi.generator' version '7.12.0'
-        id 'io.freefair.lombok' version '8.13'
     }
 }
 
-rootProject.name = 'laa-spring-boot-microservice-template'
+rootProject.name = 'laa-data-claims-reporting-service'
 
-include 'spring-boot-microservice-api'
-include 'spring-boot-microservice-service'
-
+include 'laa-data-claims-reporting-service'
