--- spring-boot-microservice-service/build.gradle
@@ -1,51 +1,0 @@
-plugins {
-    id 'io.freefair.lombok' version '8.14.2'
-    id 'uk.gov.laa.springboot.laa-spring-boot-gradle-plugin'
-}
-
-test {
-    useJUnitPlatform()
-    finalizedBy jacocoTestReport
-}
-
-jacocoTestReport {
-    afterEvaluate {
-        classDirectories.setFrom(files(classDirectories.files.collect {
-            fileTree(dir: it, exclude: ['**/SpringBootMicroserviceApplication.class'])
-        }))
-    }
-}
-
-jacocoTestCoverageVerification {
-    afterEvaluate {
-        classDirectories.setFrom(files(classDirectories.files.collect {
-            fileTree(dir: it, exclude: ['**/SpringBootMicroserviceApplication.class'])
-        }))
-    }
-}
-
-dependencies {
-    implementation project(':spring-boot-microservice-api')
-
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.12'
-
-    implementation 'org.apache.tomcat.embed:tomcat-embed-core:11.0.10'
-
-    implementation 'org.apache.commons:commons-lang3:3.18.0'
-
-    implementation 'org.mapstruct:mapstruct:1.6.3'
-    annotationProcessor 'org.mapstruct:mapstruct-processor:1.6.3'
-
-    runtimeOnly 'com.h2database:h2'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
-}
