--- spring-boot-microservice-service/src/main/resources/data.sql
@@ -1,6 +1,0 @@
--- Used to seed the example Items database
-INSERT INTO ITEMS(NAME, DESCRIPTION) VALUES('Item One', 'This is a description of Item One.');
-INSERT INTO ITEMS(NAME, DESCRIPTION) VALUES('Item Two', 'This is a description of Item Two.');
-INSERT INTO ITEMS(NAME, DESCRIPTION) VALUES('Item Three', 'This is a description of Item Three.');
-INSERT INTO ITEMS(NAME, DESCRIPTION) VALUES('Item Four', 'This is a description of Item Four.');
-INSERT INTO ITEMS(NAME, DESCRIPTION) VALUES('Item Five', 'This is a description of Item Five.');
