--- spring-boot-microservice-service/src/main/resources/schema.sql
@@ -1,7 +1,0 @@
-CREATE TABLE ITEMS
-(
-    id          BIGINT          AUTO_INCREMENT,
-    name        varchar(20)     NOT NULL,
-    description varchar(100)    NOT NULL,
-    PRIMARY KEY (id)
-);
